Fold process_file() call back into process_file_or_dir()
I initially thought that the controller would not call processors,
but it turns out that it needs to do that to, so in effect, all
callers of process_file_or_dir_with_func() were passing
process_file() to it. So let's simplify things by making
process_file_or_dir_with_func() always call process_file() again.
And then we can rename it back to the shorter name.