Rename adapters.cpython-312.fixed.pyc
When run on the tests/cases directory, we would remove "adapters.cpython-312.fixed.pyc".
This happens because we also have "adapters.cpython-312.pyc", and marshalparser
creates "adapters.cpython-312.fixed.pyc" for it, which we then remove.
Ideally, marshalparser will get an option to use a specified file name, so let's
just implement a work-around of renaming the file for now.

In practice, this is unlikely to be a problem because the suffixes of pyc files
always have the same version with ".<tag>.pyc", so we're unlikely to encounter any
.fixed.pyc files in the wild.