Allow users to undefine %add_determinism_paths or set them to %nil
When invoked as brp, %add_determinism_paths can be undefined to
disable the program. This would cause an error, because the program
checks if inputs are present.

Also, allow %add_determinism_options to be undefined.

When not running as brp, print an information message in case the user
is confused. We still instantiate the handlers, in case the user wants
to do that to check if things work.