pyc: reimplement marshalparser internally
This is a fairly straightforward reimplementation of the code in marshalparser.

The code implements (almost) full parsing of objects and prints them using
dbg!. This isn't necessary for the purpose of the code, i.e. the removal of
unused flag refs. It would be enough to read just some parts and skip over
most data, making the parser faster.

But there are good reasons to keep that part of the code:
- having the ability to print things out makes it easier to develop the code.
  I expect that adjustments will be needed for new Python versions. Also,
  we may want to add support for older Python versions, in particular Python 2.
  We have pyc files for Python 2 after all, and we want those packages to be
  reproducible too.
- I like the idea of lifting this code and having a fully functional pyc parser
  in Rust. We'd need to add the missing bits and implement proper printing.
  But at this point it seems that that wouldn't be too much work.
- The slowdown from the additional parsing and object creation doesn't seem
  to big too big.

So for now let's keep the unnecessary code. We can always drop it later.

Tested with:
$ find /usr/ -name '*.pyc'|xargs install -Dt /var/tmp/pycs1/ -m0644
$ ls -1 /var/tmp/pycs2|wc -l
45255
$ cp -a /var/tmp/pycs1/ /var/tmp/pycs2/
$ target/debug/add-determinism /var/tmp/pycs1/
$ find /var/tmp/pycs2/ -name '*.pyc' | parallel marshalparser --fix -o {}
$ diff -r /var/tmp/pycs1/ /var/tmp/pycs2/