build.rs: Use a BufWriter
I was looking at this repository recently, was curious what
the need was for a `build.rs`, and my "File open without BufWriter"
triggered.  Obviously performance is irrelevant here, but I've
had to do this multiple times in places where it *does* matter
and it's better to replicate the pattern consistently.