Add commandline option --check
In this mode, we discard any modifications that were made. If at the
end the (discarded) modification count is >0, we treat this as an
error.

The error is also returned if we were not able to parse some files.
I'm not sure what the best behaviour here is: if we are asked to check,
but the file is malformed or unsupported, the user probably wants to
know about that. But if the file just has a different type despite the
name and we fail at the initial magic test, maybe this is the wrong
thing to do. Let's do it like this for now and wait for feedback.

Resolves https://github.com/keszybz/add-determinism/issues/19.