tests: exclude pyc test files above 50kB from being published
Before this commit, the package is
    Packaged 1629 files, 29.3MiB (9.1MiB compressed)
With the change:
    Packaged 1482 files, 15.8MiB (4.2MiB compressed)

I'm on a slow network an the upload fails. Let's reduce the size of
the crate. I think this is reasonable to do in general, we don't need
so many pyc files for the downstream tests.

(create.io warns that 10MB is some sort of a limit:
> Request timed out after 30 seconds. If you're trying to upload a crate it may be too large.
> If the crate is under 10MB in size, you can email help@crates.io for assistance.
> Total size was 9498014.
I don't think we're hitting this limit yet.)