Fix supression and propagation of errors when iterating over directories
We had the following bug:
when iterating over directories, if an entry couldn't be processed,
we'd propagate the error to the caller. This means that the rest of
entries in the directory wouldn't be processed.

This was exposed by the bug that was fixed by the parent commit: we
would trip over our own temporary file, which would be briefly visible
but gone before the statx() from the .metadata() call.

Also, if the error is for the path specified on the command line,
we should fail hard, not just report an error in the statistics.
This worked in the beginning, but was broken when the rework to
collect statistics was implemented.