pyc: replace data in place
This is in preparation for future changes. The code for 'struct Pyc'
is put together in one place, separate from the more general code for
PycParser.

parser.clear_unused_flag_refs() modifies parser in place. A scratch
copy of the data is still used, so that modifications can be discarder
on failure. Returning a new copy of data would be awkward if we want
to do further processing on that data.