Remove duplicate and incomplete handling of errors
When invoked as 'add-determinism /nonexistent/path', we'd get a very verbose
error and a non-zero exit code. But when invoked as
'add-determinism -j2 /nonexistent/path', we'd get just warning message
and zero exit code. The initial idea was that a non-existent path arg would
be a fatal error, but errors in nested files would just result in a warning.
But since the introduction of the full stats, this old model doesn't make
much sense. Let's treat errors uniformly: if we fail to process a path,
warn and continue, and always exit non-zero if any errors were recorded.