Fix operation when -j with no argument is given
I completely forgot to implement the handling for -j with no args…

This was a nasty failure: we would spawn 0 workers, send jobs for any
files found, but then tell the workers to exit, which would be a noop,
and then we'd read all the results, of which there would be 0, and
happily report that now work was done. If there were enough files
found, we'd deadlock when trying to write jobs to the socket.