Do not follow symlinks when checking $RPM_BUILD_ROOT
In the check for --brp, we would canonicalize the path for $RPM_BUILD_ROOT,
and then compare the result with uncanonicalized positional args. If
RPM_BUILD_ROOT had a symlink component, we would falsely claim tha the args
are outside of $RPM_BUILD_ROOT.

The original intent was to remove duplicate slashes and such. But
std::path::Path::canonicalize looks in the file system and tries to resolve
symlinks. We actually don't want any of this, we just want to do synthetic
canonicalization without file system access.