pyc: implement a basic printer
I need to rethink how references are stored, so that it's possible to
print out the referenced items too. But this needs a restructuring
of the data, so let's do that later.

The printing is hardwired for pyc files only. I don't think the other
cases are interesting: for ar and jar files, 'ar' and 'jar' do printing.
For javadoc html files, 'less' is enough.