pyc: maximize use of references
Previous version of the cleaner only removed "unused flag references",
i.e. dropped the bit in the entry header that marked a given entry for
future reference. Multiple entries with the same type & contents could
still be present. The size of the file didn't change.

Now we go further, and make sure that a given entry is only present
once, and any future instances of the same type+contents are replaced
by a reference. This new contents should thus be either identical or
smaller than the previous form. As far as Python is concerned, this
should not result in any visible change in behaviour.

This is expected to fix issues with build reproducibility that are
still observed, in particular with ansible and other packages with
lots of python files.