pyc: fix printing of code/linetable/exceptiontable ref
We would supress printing of the binary fields when the field is a string.
But when it's a reference to a string, we'd print it in full.

This changes the output to keep the ref info, but suppress the array.
-          -code: (ref to 232)[149, 0, 91, 1, 0, 0, 0, 0, 0, 0, 0, 0, 85, 1, 53, 1, 0, 0, 0, 0, 0, 0, 110, 1, 85, 1, 91, 2, 0, 0, 0, 0, 0, 0, 0, 0, 76, 0, 97, 2, 0, 0, 85, 1, 36, 0, 85, 1, 82, 5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 88, 2, 83, 1, 57, 2, 36, 0]
+          -code: (ref to 232)[76 bytes]

The formatting is changed to use brackets, to make the output more
similar to the array. When array is empty, just "[]" is shown, which
is shorter than "[0 bytes]".