tests/pyc: rewrite all "fixed" files with the new handler
Originally, "fixed" files were produced using marshalparser, so we
could be fairly confident that they are correct. But now we're
departing from what marshalparser was doing, so we cannot compare the
new output with anything except itself.

To verify that the output is good, I first applied the handler to all
.pyc files, and then compared the print output for the .pyc files
with the output for all the .pyc.fixed files, suppressing reference
info:

$ for f in $(git ls '*.pyc.fixed'|sed 's/\.fixed$//'); do
      r=target/debug/add-determinism
      echo $f
      diff -u <($r -p $f | sed -r 's/\(ref to [0-9]+\)//g; s/ \[#[0-9]+\]//g') \
              <($r -p $f.fixed | sed -r 's/\(ref to [0-9]+\)//g; s/ \[#[0-9]+\]//g') | colordiff
      echo
  done

The output appears to be the same. The only difference is that the
printer opportinistically prints tuples in one line, and changing the
reference flags influences this. So the outputs are not identical, but
almost so.

Copy&paste error in test name is fixed.