pyc: stop replacing PyLong/'l' by Int/'i'
This was incomplete, because the hash and equality functions didn't take this
into account, resulting in unstable output (the first run would write 'i' with
the same value twice, once from an original PyLong object and once from an
original Int object, but the rerun would see two Ints, and replace the second
by 'r').

Let's remove this for now. We can add it back more systematically at a later
point.