pyc: print a warning if trailing data is found
This is very much unexpected. Our screwed-up code did that for a while, so
obviously there will be _some_ files like this out in the wild. The good news
is that nothing seems to care about this. In fact, all three printers didn't
even mention the trailing data and cpython doesn't complain either.

We don't have a good mechanism to convery warnings but to the main program.
But a warning in the logs should be enough.