pyc: fix the header in rewritten pyc files
I managed to somehow get this completely wrong. That number is the size of the
original .py file, nothing related to the bytecode file itself. Unfortunately
this causes python to ignore the pyc file (if the py file is there to
compare), so it's a noticable bug. The packages that were built with that will
need to be rebuilt.

Unfortunately the test files that were adjusted previously need to be adjusted
again.