pyc: fix handling of refences to objects
When the original file had the same object twice, once normally and
once as a reference, when writing out, we'd write the object a second
time in full, because we had Ref(x)!=y, even when x==y. We handled
this correctly in the hashing code, and the writing code, but not in
the comparison function. The fix is easy — just deference in one more
place, but it took me forever to figure out.