pyc: adjust TODOs
The "content hash" is for the _source_ file, so it doesn't change if we
rewrite the pyc file. The TODO entry was completely bogus.

But as a follow-up, it'd be nice if we could notice the cases where we're
working on a file that doesn't actually work for the size and/or mtime of the
pyc file. This could help to diagnose the cases where the py file mtime was
changed after the pyc file was written.