pyc: rename "ref" to "flag"
The old naming was copied from marshalparser. But it mixes the concept of a
"flag", i.e. the setting of the flag bit and keeping a list of flagged items,
with references to those items. So let's rename to make this all easier to
grok.

In particular, format_flag(show_flag, flag_num) is much better than
show_ref(show_ref, ref_index).