handlers: always set output_path in InputOutputHelper
SOURCE_DATE_EPOCH=1748901600 target/debug/add-determinism --check tests/cases/ -j
would fail an unwrap in zip handler because because io.output_path is None, and it
tries to reopen the output path. I think this wasn't caught before because both
"--check" and appropriate "$SOURCE_DATE_EPOCH" are required.

Let's rework the helper to always set .output_path and instead use .check as the
flag whether the output file is fake. I think this is better.

The zip handler will still fail, because it tries to re-read the output file as
a zip file, which will not work with --check beause /dev/null does not have a
valid header.