Reorganize tree structure
The stuff for the add-determinism binary is moved one level down, to the
src/add_det subdirectory. (Yes, I think it's time to rename the binary to just
'add-det'. Typing the full name is annoying.)

The tests are rearranged a bit, with the shared code moved to
tests/common/mod.rs. This way cargo test doesn't try to build it into a
separate test binary, but it can still be imported from anywhere.
Unfortunately, a tests/tests.rs file is still needed, because otherwise
test/handlers/mod.rs isn't run.