linkdupes: check SELinux context of files that could be linked
If we figure out that two files are candidates for linking, check what
the SELinux policy says their contexts would be. If they are different,
do not link them. This is pretty much the same as other metadata:
we need the same mtime, owner, group, mode, and context, if the file
is to be linked.

The implementation is not complete, because the file *type* is not
passed in the look_up_by_path call. For rpm packaging, this doesn't
really matter, so I'm leaving that for later.