linkdupes: make the sort order stable
When running the program, the printed output would differ if the sorting
algorithm changed (I was expemerimenting with rayon::par_sort_unstable_by).
This makes comparing different runs more difficult.

The order should be stable, in particular it should not depend on the sort
algorithm and it should not depend on the directory entry order. So when
sorting, let's compare all entries as unequal and as a last resort compare the
file names. Later when linking, we obviously should not do that, so that we
can actually link files with different names.