wip: use rayon to paralellize sorting
This does not work (easily) with selinux support, because
trait Sync is not implemented for NonNull<selinux_sys::selabel_handle>, and we get:
error[E0277]: NonNull<selinux_sys::selabel_handle> cannot be shared between threads safely

But without selinux support, this compiles and runs nicely. But there
is no gain or very little gain in wallclock time from the parallel
version. On my 4-core/8-thread machine, user CPU use goes from 99% to
394%. Occasionally the wallclock time is a few percent smaller, but
sometimes it is 2-3 times higher. (Maybe something with how the threads
are scheduled.)  This does not seem to make sense to use the parallelized
version.